refactor: split explore web tab
